fn comments; def params

+++ LEAF_Request_Portal/admin/templates/mod_svcChief.tpl
@@ -47,6 +47,9 @@
 <script type="text/javascript">
 /* <![CDATA[ */
 
+/**
+ * Sync current list of services to that in the Nexus.
+ */
 function syncServices() {
     dialog_simple.setTitle('Importing from Nexus...');
     dialog_simple.show();
@@ -58,7 +61,7 @@ function syncServices() {
         },
         fail: function(error) {
             console.log(error);
-        }
+        },
         cache: false
     });
     dialog_simple.setCancelHandler(function() {
@@ -66,6 +69,9 @@ function syncServices() {
     });
 }
 
+/**
+ * Create new service. [DEPRECATED]
+ */
 function createGroup() {
 	/*
 	dialog.clear();
@@ -110,7 +116,11 @@ function createGroup() {
     dialog_simple.show();
 }
 
-function getMembers(groupID) {
+/**
+ * Get members for given service.
+ * @param {int} groupID - ID for group
+ */
+function getMembers(groupID = -1) {
     $.ajax({
         type: 'GET',
         url: '../api/system/updateService/' + groupID,
@@ -125,7 +135,7 @@ function getMembers(groupID) {
                 },
                 fail: function(error) {
                     console.log(error);
-                }
+                },
                 cache: false
             });
         },
@@ -133,7 +143,15 @@ function getMembers(groupID) {
     });
 }
 
-function populateMembers(groupID, members) {
+/**
+ * Populate the members of a given service.
+ * @param {int} groupID - ID of given group
+ * @param {array} members - list object of all members in group
+ */
+function populateMembers(groupID = -1, members = []) {
+    if (groupID < 0 || members.length === 0) {
+        return;
+    }
     $('#members' + groupID).html('');
     let memberCt = -1;
     for (let i in members) {
@@ -154,39 +172,53 @@ function populateMembers(groupID, members) {
 
 /**
  * Add user to portal
- * @param groupID - ID of group
- * @param userID - ID of user being added
+ * @param {int} groupID - ID of group
+ * @param {int} userID - ID of user being added
  */
-function addUser(groupID, userID) {
-    $.ajax({
-        type: 'POST',
-        url: "../api/service/" + groupID + "/members",
-        data: {'userID': userID,
-               'CSRFToken': '<!--{$CSRFToken}-->'},
-        cache: false
-    });
+function addUser(groupID = -1, userID = -1) {
+    if (groupID < 0 || userID < 0) {
+        return;
+    } else {
+        $.ajax({
+            type: 'POST',
+            url: "../api/service/" + groupID + "/members",
+            data: {'userID': userID,
+                'CSRFToken': '<!--{$CSRFToken}-->'},
+            cache: false
+        });
+    }
 }
 
 /**
  * Remove user from portal.
- * @param groupID - ID of group
- * @param userID - ID of user being removed
+ * @param {int} groupID - ID of group
+ * @param {int} userID - ID of user being removed
  */
-function removeUser(groupID, userID) {
-    $.ajax({
-        type: 'DELETE',
-        url: "../api/service/" + groupID + "/members/_" + userID + '?' +
-            $.param({'CSRFToken': '<!--{$CSRFToken}-->'}),
-        cache: false
-    });
+function removeUser(groupID = -1, userID = -1) {
+    if (groupID < 0 || userID < 0) {
+        return;
+    } else {
+        $.ajax({
+            type: 'DELETE',
+            url: "../api/service/" + groupID + "/members/_" + userID + '?' +
+                $.param({'CSRFToken': '<!--{$CSRFToken}-->'}),
+            cache: false
+        });
+    }
 }
 
 /**
  * Import user from orgchart.
- * @param serviceID - ID of service
- * @param selectedUserName - Username being imported
+ * @param {int} serviceID - ID of service
+ * @param {string} selectedUserName - Username being imported
  */ 
-function importUser(serviceID, selectedUserName) {
+function importUser(serviceID = 0, selectedUserName = '') {
+    if (serviceID === 0) {
+        console.log('Invalid serviceID');
+    }
+    if (selectedUserName === '') {
+        console.log('Invalid username');
+    }
     $.ajax({
         type: 'POST',
         url: '<!--{$orgchartPath}-->/api/employee/import/_' + selectedUserName,
@@ -208,114 +240,129 @@ function importUser(serviceID, selectedUserName) {
 
 /**
  * Delete given service.
- * @param int serviceID - ID of the service
+ * @param {int} serviceID - ID of the service
  */
-function deleteService(serviceID) {
-    $.ajax({
-        type: 'DELETE',
-        url: "../api/service/" + serviceID + '?' +
-            $.param({'CSRFToken': '<!--{$CSRFToken}-->'}),
-        success: function(response) {
-            location.reload();
-        },
-        fail: function(error) {
-            console.log(error);
-        },
-        cache: false
-    });
+function deleteService(serviceID = 0) {
+    if (serviceID === 0) {
+        return;
+    } else {
+        $.ajax({
+            type: 'DELETE',
+            url: "../api/service/" + serviceID + '?' +
+                $.param({'CSRFToken': '<!--{$CSRFToken}-->'}),
+            success: function(response) {
+                location.reload();
+            },
+            fail: function(error) {
+                console.log(error);
+            },
+            cache: false
+        });
+    }
 }
 
 /**
  * Build the modal for when the user selects a service.
- * @param int serviceID - ID of service
- * @param string serviceName - Name of the service
+ * @param {int} serviceID - ID of service
+ * @param {string} serviceName - Name of the service
  */ 
-function initiateModal(serviceID, serviceName) {
-    $.ajax({
-        type: 'GET',
-        url: '../api/service/' + serviceID + '/members',
-        success: function(res) {
-            dialog.clear();
-            let button_deleteGroup = '<div><button id="deleteGroup_'+serviceID+'" class="usa-button usa-button--secondary leaf-btn-small leaf-marginTop-1rem">Delete Group</button></div>';
-            if(serviceID > 0) {
-                button_deleteGroup = '';
-            }
-            dialog.setContent(
-                '<div class="leaf-float-right"><div><button class="usa-button leaf-btn-small" onclick="viewHistory('+serviceID+')">View History</button></div></div>' +
-                '<a class="leaf-group-link" href="<!--{$orgchartPath}-->/?a=view_group&groupID=' + serviceID + '" title="groupID: ' + serviceID + '" target="_blank"><h2 role="heading" tabindex="-1">' + serviceName + '</h2></a><h3 role="heading" tabindex="-1" class="leaf-marginTop-1rem">Add Employee</h3><div id="employeeSelector"></div></br><div id="employees"></div>');
-            $('#employees').html('<div id="employee_table" class="leaf-marginTopBot-1rem"></div>');
-            let counter = 0;
-            for(let i in res) {
-                // Check for active members to list
-                if (res[i].active == 1) {
-                    if (res[i].backupID == null) {
-                        let removeButton = '- <a href="#" class="text-secondary-darker leaf-font0-7rem leaf-remove-button" id="removeMember_' + counter + '">REMOVE</a>';
-                        $('#employee_table').append('<a href="<!--{$orgchartPath}-->/?a=view_employee&empUID=' + res[i].empUID + '" class="leaf-user-link" title="' + res[i].empUID + ' - ' + res[i].userName + '" target="_blank"><div class="leaf-marginTop-halfRem leaf-bold leaf-font0-9rem">' + toTitleCase(res[i].Lname) + ', ' + toTitleCase(res[i].Fname) + '</a> <span class="leaf-font-normal">' + removeButton + '</span></div>');
-                        // Check for Backups
-                        for (let j in res) {
-                            if (res[i].userName == res[j].backupID) {
-                                $('#employee_table').append('<div class="leaf-font0-8rem leaf-marginLeft-qtrRem">&bull; ' + toTitleCase(res[j].Fname) + ' ' + toTitleCase(res[j].Lname) + ' - <span class="text-secondary-darker leaf-font0-7rem">Backup for ' + toTitleCase(res[i].Fname) + ' ' + toTitleCase(res[i].Lname) + '</span></div>');
+function initiateModal(serviceID = 0, serviceName = '') {
+    if (serviceID === 0 || serviceName === '') {
+        return;
+    } else {
+        $.ajax({
+            type: 'GET',
+            url: '../api/service/' + serviceID + '/members',
+            success: function(res) {
+                dialog.clear();
+                let button_deleteGroup = '<div><button id="deleteGroup_'+serviceID+'" class="usa-button usa-button--secondary leaf-btn-small leaf-marginTop-1rem">Delete Group</button></div>';
+                if(serviceID > 0) {
+                    button_deleteGroup = '';
+                }
+                dialog.setContent(
+                    '<div class="leaf-float-right"><div><button class="usa-button leaf-btn-small" onclick="viewHistory('+serviceID+')">View History</button></div></div>' +
+                    '<a class="leaf-group-link" href="<!--{$orgchartPath}-->/?a=view_group&groupID=' + serviceID + '" title="groupID: ' + serviceID + '" target="_blank"><h2 role="heading" tabindex="-1">' + serviceName + '</h2></a><h3 role="heading" tabindex="-1" class="leaf-marginTop-1rem">Add Employee</h3><div id="employeeSelector"></div></br><div id="employees"></div>');
+                $('#employees').html('<div id="employee_table" class="leaf-marginTopBot-1rem"></div>');
+                let counter = 0;
+                for(let i in res) {
+                    // Check for active members to list
+                    if (res[i].active == 1) {
+                        if (res[i].backupID == null) {
+                            let removeButton = '- <a href="#" class="text-secondary-darker leaf-font0-7rem leaf-remove-button" id="removeMember_' + counter + '">REMOVE</a>';
+                            $('#employee_table').append('<a href="<!--{$orgchartPath}-->/?a=view_employee&empUID=' + res[i].empUID + '" class="leaf-user-link" title="' + res[i].empUID + ' - ' + res[i].userName + '" target="_blank"><div class="leaf-marginTop-halfRem leaf-bold leaf-font0-9rem">' + toTitleCase(res[i].Lname) + ', ' + toTitleCase(res[i].Fname) + '</a> <span class="leaf-font-normal">' + removeButton + '</span></div>');
+                            // Check for Backups
+                            for (let j in res) {
+                                if (res[i].userName == res[j].backupID) {
+                                    $('#employee_table').append('<div class="leaf-font0-8rem leaf-marginLeft-qtrRem">&bull; ' + toTitleCase(res[j].Fname) + ' ' + toTitleCase(res[j].Lname) + ' - <span class="text-secondary-darker leaf-font0-7rem">Backup for ' + toTitleCase(res[i].Fname) + ' ' + toTitleCase(res[i].Lname) + '</span></div>');
+                                }
                             }
+                            $('#removeMember_' + counter).on('click', function (userID) {
+                                return function () {
+                                    removeUser(serviceID, userID);
+                                    dialog.hide();
+                                };
+                            }(res[i].userName));
+                            counter++;
                         }
-                        $('#removeMember_' + counter).on('click', function (userID) {
-                            return function () {
-                                removeUser(serviceID, userID);
-                                dialog.hide();
-                            };
-                        }(res[i].userName));
-                        counter++;
                     }
                 }
-            }
 
-            $('#deleteGroup_' + serviceID).on('click', function() {
-                dialog_confirm.setContent('Are you sure you want to delete this service?');
-                dialog_confirm.setSaveHandler(function() {
-                    deleteService(serviceID);
+                $('#deleteGroup_' + serviceID).on('click', function() {
+                    dialog_confirm.setContent('Are you sure you want to delete this service?');
+                    dialog_confirm.setSaveHandler(function() {
+                        deleteService(serviceID);
+                    });
+                    dialog_confirm.show();
                 });
-                dialog_confirm.show();
-            });
 
-            empSel = new nationalEmployeeSelector('employeeSelector');
-            empSel.apiPath = '<!--{$orgchartPath}-->/api/?a=';
-            empSel.rootPath = '<!--{$orgchartPath}-->/';
-            empSel.outputStyle = 'micro';
-            empSel.initialize();
-            // Update on any action
-            dialog.setCancelHandler(function() {
-                getMembers(serviceID);
-            });
-            dialog.setSaveHandler(function() {
-                if(empSel.selection != '') {
-                    let selectedUserName = empSel.selectionData[empSel.selection].userName;
-                    importUser(serviceID, selectedUserName);
-                }
-                dialog.hide();
-            });
+                empSel = new nationalEmployeeSelector('employeeSelector');
+                empSel.apiPath = '<!--{$orgchartPath}-->/api/?a=';
+                empSel.rootPath = '<!--{$orgchartPath}-->/';
+                empSel.outputStyle = 'micro';
+                empSel.initialize();
+                // Update on any action
+                dialog.setCancelHandler(function() {
+                    getMembers(serviceID);
+                });
+                dialog.setSaveHandler(function() {
+                    if(empSel.selection != '') {
+                        let selectedUserName = empSel.selectionData[empSel.selection].userName;
+                        importUser(serviceID, selectedUserName);
+                    }
+                    dialog.hide();
+                });
 
-            dialog.show();
-        },
-        fail: function(error) {
-            console.log(error);
-        },
-        cache: false
-    });
+                dialog.show();
+            },
+            fail: function(error) {
+                console.log(error);
+            },
+            cache: false
+        });
+    }
 }
 
 /**
  * Initiate widgets for each service 
- * @param serviceID - ID for service being represented
- * @param serviceName - name of service being represented
+ * @param {int} serviceID - ID for service being represented
+ * @param {string} serviceName - name of service being represented
  * @return function
  */
-function initiateWidget(serviceID, serviceName) {
-    $('#' + serviceID).on('click', function(serviceID) {
-        return function() {
-            initiateModal(serviceID, serviceName);
-        };
-    }(serviceID));
+function initiateWidget(serviceID = 0, serviceName = '') {
+    if (serviceID === 0 || serviceName === '') {
+        return;
+    } else {
+        $('#' + serviceID).on('click', function(serviceID) {
+            return function() {
+                initiateModal(serviceID, serviceName);
+            };
+        }(serviceID));
+    }
 }
 
+/**
+ * Get list of present services in portal.
+ */
 function getGroupList() {
 	$.when(
 	    $.ajax({
@@ -346,7 +393,14 @@ function getGroupList() {
 	});
 }
 
-function viewHistory(groupID){
+/**
+ * View history of given group by ID.
+ * @param {int} groupID - ID of given group
+ */
+function viewHistory(groupID = -1){
+    if (groupID < 0) {
+        return;
+    }
     dialog_simple.setContent('');
     dialog_simple.setTitle('Service chief history');
 	dialog_simple.show();
@@ -362,13 +416,16 @@ function viewHistory(groupID){
         },
         fail: function(error) {
             console.log(error);
-        }
+        },
         cache: false
     });
 }
 
-// convert to title case
-function toTitleCase(str) {
+/** 
+ * Convert to title case.
+ * @param {string} str - string to be converted
+ */
+function toTitleCase(str = '') {
     return (str == "" || str == null) ? "" : str.replace(/\w\S*/g, function(txt){return txt.charAt(0).toUpperCase() + txt.substr(1).toLowerCase();});
 }
